[Add] DeleteTrackTests - only_authorized_users_can_delete_his_tracks

+++ app/Http/Controllers/TrackController.php
@@ -28,6 +28,8 @@ class TrackController extends Controller
 
     public function destroy(Track $track)
     {
+        $this->authorize('delete', $track);
+
         $track->delete();
 
         return response()->json()->setStatusCode(204);
